Fix bug of default post date not updated
Originally the post schema date
property's default value is assigned with
a Date object representing the time when the schema is created. This
changes this default property value to a arrow function that returns a
newly created Date object, as a result, each time a blog post is
created, the post date value would be the date when the post is created
instead of the date when the schema is created.